Move function to Cardano.Api.Extras
